Ensure integer wrapping instead of preventing overflow/underflow

+++ src/engine/shared/snapshot.h
@@ -89,18 +89,18 @@ private:
 	};
 	short m_aItemSizes[MAX_NETOBJSIZES];
 	short m_aItemSizes7[MAX_NETOBJSIZES];
-	int m_aSnapshotDataRate[CSnapshot::MAX_TYPE + 1];
-	int m_aSnapshotDataUpdates[CSnapshot::MAX_TYPE + 1];
+	uint64_t m_aSnapshotDataRate[CSnapshot::MAX_TYPE + 1];
+	uint64_t m_aSnapshotDataUpdates[CSnapshot::MAX_TYPE + 1];
 	CData m_Empty;
 
-	static void UndiffItem(const int *pPast, const int *pDiff, int *pOut, int Size, int *pDataRate);
+	static void UndiffItem(const int *pPast, const int *pDiff, int *pOut, int Size, uint64_t *pDataRate);
 
 public:
 	static int DiffItem(const int *pPast, const int *pCurrent, int *pOut, int Size);
 	CSnapshotDelta();
 	CSnapshotDelta(const CSnapshotDelta &Old);
-	int GetDataRate(int Index) const { return m_aSnapshotDataRate[Index]; }
-	int GetDataUpdates(int Index) const { return m_aSnapshotDataUpdates[Index]; }
+	uint64_t GetDataRate(int Index) const { return m_aSnapshotDataRate[Index]; }
+	uint64_t GetDataUpdates(int Index) const { return m_aSnapshotDataUpdates[Index]; }
 	void SetStaticsize(int ItemType, size_t Size);
 	void SetStaticsize7(int ItemType, size_t Size);
 	const CData *EmptyDelta() const;
